Move globals and helper function up to the test subpackage def
This allows them to more easily be used in other test modules.
