+++ permafrost_benchmark_system/tests/test_modelingest.py
@@ -2,24 +2,10 @@
 
 import os
 import shutil
-import yaml
 from nose.tools import nottest, assert_true, assert_equal
 from permafrost_benchmark_system.ingest import ModelIngest
-
-
-ingest_file = 'test_ingest.yaml'
-model_file = 'test_model.txt'
-note_file = model_file + '.txt'
-tmp_dir = 'tmp'
-
-
-@nottest
-def make_test_files():
-    with open(model_file, 'w') as fp:
-        fp.write('This is a test model output file.\n')
-    cfg = {'ingest_files': [model_file], 'make_public': True}
-    with open(ingest_file, 'w') as fp:
-        yaml.safe_dump(cfg, fp, default_flow_style=False)
+from . import (ingest_file, model_file, note_file, tmp_dir,
+               make_test_files)
 
 
 def setup_module():
